Add dummy example using sessions.
The example shows the basic usage of sessions to store the userdata.
The sessions will be stored using the secret in the clients cache.

Unless the user figures out the secret it will not be easy to tamper with the session and therefore should be quite save to use in that way.